requirements.txt
pyperclip
pyyaml
spacy
presidio-analyzer
presidio-anonymizer
keyboard
winotify
pystray
Pillow
